+++ src/app/createAlbaform/components/Datepicker.styles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { FormInput } from './StepForm.styles';
+import { FormInput } from './Form.styles';
 
 //  달력 커스텀 인풋
 export const CustomDateInput = styled(FormInput)`
